requirements.txt
matplotlib>=3.4.2
numpy>=1.21.1
opencv-python>=4.5.5.64
Pillow>=9.2.0
tensorboard>=2.5.0
torch>=2.0.1
torchvision>=0.15.2
pytorch-lightning>=1.6.5
torchmetrics>=0.11.3
fire>=0.5.0
loguru>=0.6.0
face-alignment>=1.3.5
scikit-image>=0.19.3
scikit-learn>=1.2.0
PySimpleGUI>=4.60.4
mediapipe>=0.9.0.1
pandas>=1.5.2
segmentation-models-pytorch>=0.3.2
seaborn>=0.12.2
deepface>=0.0.79
retina-face>=0.0.13
tqdm>=4.65.0
p_tqdm>=1.4.0
statsmodels==0.13.5
scipy>=1.10.1
cmake>=3.26.1
dlib>=19.24.1
git+https://github.com/wiktorlazarski/head-segmentation@main